Create initial README.md
This is the very first commit on behalf of the initiative to provide a starter kit for Swiss Rudimental Drumming notation using SMuFL standard glyphs, which are available through fonts such as Bravura in applications like MuseScore and Dorico.

This repository is a response to the support we have received from the SMuFL community, the MuseScore development team, and the program management of Dorico to implement what was initiated with issue #118 (https://github.com/w3c/smufl/issues/118).

The current initiative is driven by contributions from Benjamin Fassbind, Maurice Weiss, Philipp Rütsche, and the initiator, Urs Gehrig.

The starter kit aims to provide at least a two-solution strategy with MuseScore and Dorico, offering digitized notation across multiple platforms such as Windows, macOS, and Linux. It will assist users in configuring these solutions to quickly benefit from the tools and to build upon templates for typical rudimental drumming styles, such as marches, Basel march, and solo duo, among others. While notation is just the beginning of an end-to-end workflow in music production, it is also important to address MIDI integration and post-processing later on.

Basel (Switzerland), November 25, 2024.